Added RTL for Symobi. Symobi uses the NetBSD C-library, so the RTL was started from the FreeBSD RTL (as I don't trust the state of the NetBSD RTL ^^) and adjusted accordingly.
The changes in Unix are mostly that "linklib foo" should not be used on Symobi as imported DLLs are taken care of similar to Windows. Also some functions are not available in Symobi's C-library (or not working correctly).

Please note: As most files in rtl/symobi are copied from the BSD RTLs the copyright and file information might not be correct.

* compiler/systems/i_symobi.pas
   Forgot to add the BSD define
* rtl/unix/oscdecl.inc
   Symobi does currently not define SetTimeOfDay (might be, because it's not even defined by POSIX...)
* rtl/unix/sysheap.inc
   (un)mmap does not work correctly
* rtl/inc/fpextres.pp:
   use the generic external implementation for Symobi


git-svn-id: http://svn.freepascal.org/svn/fpc/branches/svenbarth@18061 3ad0048d-3df7-0310-abae-a5850022a9f2